Add dynamic grid functionality
Created a prompt which prompts the user for the number of rows they wish
to have in their grid. Updated the createGrid function to be dynamic
instead of the default 16x16 grid. I want to limit the max number of
rows to 100 to prevent performance issues but will add that
functionality at a later point.