Modify user input to only be between 1 and 100
Created a validateInput function which checks that the input enter when
prompted returns only a number between 1 and 100, otherwise it will
continue to prompt the user for a valid input. This is so the user does
not input negative numbers, strings, special characters, or a number too
large which would break the program.